Remove overwritten padding for active nav links
Currently clicking a nav item causes its height to change, this is especially noticeable in the mobile-style layout with the app drawer, used when the browser width is very small, but also affects other layouts. It causes useability issues, making links hard to click, occasionally causing the wrong link to be clicked. This fixes that.